+++ tracer/test/Datadog.Trace.Security.IntegrationTests/IAST/AspNetCore2IastTests.cs
@@ -1,9 +1,9 @@
-// <copyright file="AspNetCore5IastTests.cs" company="Datadog">
+// <copyright file="AspNetCore2IastTests.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
-#if NETCOREAPP3_0_OR_GREATER
+#if NETCOREAPP2_1
 
 using System;
 using System.Text.RegularExpressions;
@@ -12,14 +12,14 @@ using Datadog.Trace.TestHelpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Datadog.Trace.Security.IntegrationTests
+namespace Datadog.Trace.Security.IntegrationTests.Iast
 {
-    public class AspNetCore5IastTests : AspNetBase, IDisposable
+    public class AspNetCore2IastTests : AspNetBase, IDisposable
     {
         private static readonly Regex LocationMsgRegex = new(@"(\S)*""location"": {(\r|\n){1,2}(.*(\r|\n){1,2}){0,3}(\s)*},");
 
-        public AspNetCore5IastTests(ITestOutputHelper outputHelper)
-            : base("AspNetCore5", outputHelper, "/shutdown", testName: nameof(AspNetCore5IastTests))
+        public AspNetCore2IastTests(ITestOutputHelper outputHelper)
+            : base("AspNetCore2", outputHelper, "/shutdown", testName: nameof(AspNetCore2IastTests))
         {
         }
 
@@ -27,9 +27,9 @@ namespace Datadog.Trace.Security.IntegrationTests
         [InlineData(true)]
         [InlineData(false)]
         [Trait("RunOnWindows", "True")]
-        public async Task TestIastWeakHashing(bool enableIast)
+        public async Task TestIastWeakHashingRequest(bool enableIast)
         {
-            var filename = enableIast ? "weakHashing.AspNet5.IastEnabled" : "weakHashing.AspNet5.IastDisabled";
+            var filename = enableIast ? "weakHashing.AspNetCore2.IastEnabled" : "weakHashing.AspNetCore2.IastDisabled";
             var url = "/Iast/WeakHashing";
             var agent = await RunOnSelfHosted(enableSecurity: false, enableIast: enableIast);
             var spans = await SendRequestsAsync(agent, new string[] { url });
